Fix start client script

+++ start_client.sh
@@ -1,4 +1,17 @@
 #!/bin/bash
 
-cp .tally_cache .tally_cache_client
+TALLY_CACHE=.tally_cache
+TALLY_CACHE_CLIENT=.tally_cache_client
+
+# Copy cache if not already exists
+if [ -e "$TALLY_CACHE" ]; then
+    if [ -e "$TALLY_CACHE_CLIENT" ]; then
+        if ! cmp -s "$TALLY_CACHE" "$TALLY_CACHE_CLIENT"; then
+            cp .tally_cache .tally_cache_client
+        fi
+    else
+        cp .tally_cache .tally_cache_client
+    fi
+fi
+
 LD_PRELOAD=./build/libtally_client.so $@
